+++ docs/src/markdownDocs/workflowDocs/Example/constants/index.ts
@@ -1,0 +1,3 @@
+export const USER_SETTINGS = 'user_settings';
+export const LOCAL_USER_DATA = 'user_data';
+export const REMEMBER_ME_DATA = 'remember_me_data';